Add channel priority disclaimer
This PR adds a channel priority disclaimer which indicates that `channel_priority: strict` is not supported with `conda` installs.

Closes #218.

Additionally, this makes some minor changes to the HTML structure of notes and adds some margin to the bottom of each note since it seems that we're adding quite a few now.