Update existing redirects
This PR is a continuation of #275.

Since we've recently switched our site hosting to Netlify, and therefore have access to server side redirects via the `_redirects` file, we can remove the `jekyll-redirect-from` plugin (which does client side redirects) and replace any instances where it was used with proper server side redirects.

This PR makes those changes.