mv root `index.html` to layout; mv partners & logos/metadata to new `_partners` content type; create `_partners` archetype
the benefits of moving `index.html` to `layouts` is that we can use Hugo's templating system to generate the page, and we can use Hugo's image processing functionality to resize the logos to appropriate sizes.
additionally, this change provides a clearer separation between content and layouts.

the benefit of moving partner logos and metadata to a new `_partners` content type is that the resources are well organized and easy to find and maintain.

the benefit of creating a `_partners` archetype is that it provides a template for creating new partner pages.