Fix headless `_partners` bundle
The `_partners` bundle is intended to be headless, but it currently shows up on our sitemap.

This PR includes some changes to ensure that the `_partners` bundle is actually headless.

It updates some frontmatter and the Hugo version. The older Hugo version was not respecting the headless frontmatter.

Note, this doesn't update Hugo to the latest version since that also requires updating our theme, Docsy, which is a bit more involved. That'll come later.